feat: implement persistent shell sessions for terminal state
- Add persistent session support to BashExecutor and ZshExecutor
- Environment variables now persist across commands
- Working directory tracked and maintained automatically
- Uses stdin/stdout communication with unique markers for output parsing
- Thread-safe with asyncio.Lock
